Upgrade innoextract to v1.6

+++ App/jni/innoextract/src/stream/block.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2013 Daniel Scharrer
+ * Copyright (C) 2011-2016 Daniel Scharrer
  *
  * This software is provided 'as-is', without any express or implied
  * warranty.  In no event will the author(s) be held liable for any damages
@@ -42,7 +42,7 @@
 #include "util/enum.hpp"
 #include "util/load.hpp"
 #include "util/log.hpp"
-#include "util/util.hpp"
+#include "util/math.hpp"
 
 namespace io = boost::iostreams;
 
@@ -78,7 +78,7 @@ public:
 	
 	inno_block_filter() : pos(0), length(0) { }
 	
-	template<typename Source>
+	template <typename Source>
 	bool read_chunk(Source & src) {
 		
 		char temp[sizeof(boost::uint32_t)];
@@ -86,7 +86,7 @@ public:
 		std::streamsize nread = boost::iostreams::read(src, temp, temp_size);
 		if(nread == EOF) {
 			return false;
-		} else if(nread != sizeof(temp)) {
+		} else if(size_t(nread) != sizeof(temp)) {
 			throw block_error("unexpected block end");
 		}
 		boost::uint32_t block_crc32 = util::little_endian::load<boost::uint32_t>(temp);
@@ -108,7 +108,7 @@ public:
 		return true;
 	}
 	
-	template<typename Source>
+	template <typename Source>
 	std::streamsize read(Source & src, char * dest, std::streamsize n) {
 		
 		std::streamsize read = 0;
